fix point

+++ app/src/main/java/com/example/testtaskintegrio/presenter/model/Point.kt
@@ -8,5 +8,5 @@ data class Point(
     val corpus: Long,
     val geoHash: String,
     val coordinates: GeoPoint,
-    val properties: Map<String, Int>
+    val properties: Map<String, Long>
 )
